requirements.txt
PyOpenGL
numpy
glfw
PyOpenGL_accelerate